add basic MPEG1/2 decoder (alpha)
This is the first working alpha version with many unhandled cases
which might lead to unexpected results or crashes.
Most of the code is only for handling and reordering framebuffers, the
decoding is pretty simple and short. But it still uses some magic
constants where i didn't figure out what they mean.